Resolved suggestions
