+++ ModernWpf/IconSource/IconSource.cs
@@ -1,17 +1,28 @@
 ﻿using System.Windows;
 using System.Windows.Media;
-
 namespace ModernWpf.Controls
 {
+    /// <summary>
+    /// Represents the base class for an icon source.
+    /// </summary>
     public class IconSource : DependencyObject
     {
+        /// <summary>
+        /// Identifies the <see cref="Foreground"/> dependency property.
+        /// </summary>
         public static readonly DependencyProperty ForegroundProperty =
             DependencyProperty.Register(
                 nameof(Foreground),
                 typeof(Brush),
-                typeof(IconSource),
-                new PropertyMetadata(null));
+                typeof(IconSource));
 
+        /// <summary>
+        /// Gets or sets a brush that describes the foreground color.
+        /// </summary>
+        /// <returns>
+        /// The brush that paints the foreground of the control. The default is <see langword="null"/>, (a null brush) which is
+        /// evaluated as Transparent for rendering.
+        /// </returns>
         public Brush Foreground
         {
             get => (Brush)GetValue(ForegroundProperty);
