+++ ModernWpf/IconSource/SymbolIconSource.cs
@@ -2,23 +2,38 @@
 
 namespace ModernWpf.Controls
 {
+    /// <summary>
+    /// Represents an icon source that uses a glyph from the Segoe MDL2 Assets font as its content.
+    /// </summary>
     public class SymbolIconSource : IconSource
     {
-		public SymbolIconSource()
-		{
-		}
+        /// <summary>
+        /// Initializes a new instance of the <see cref="SymbolIconSource"/> class.
+        /// </summary>
+        public SymbolIconSource()
+        {
+        }
 
-		public static readonly DependencyProperty SymbolProperty =
-			DependencyProperty.Register(
-				nameof(Symbol),
-				typeof(Symbol),
-				typeof(SymbolIconSource),
-				new PropertyMetadata(Symbol.Emoji));
+        /// <summary>
+        /// Identifies the <see cref="Symbol"/> dependency property.
+        /// </summary>
+        public static readonly DependencyProperty SymbolProperty =
+            DependencyProperty.Register(
+                nameof(Symbol),
+                typeof(Symbol),
+                typeof(SymbolIconSource),
+                new PropertyMetadata(Symbol.Emoji));
 
-		public Symbol Symbol
-		{
-			get => (Symbol)GetValue(SymbolProperty);
-			set => SetValue(SymbolProperty, value);
-		}
-	}
+        /// <summary>
+        /// Gets or sets the Segoe MDL2 Assets glyph used as the icon content.
+        /// </summary>
+        /// <returns>
+        /// A named constant of the enumeration that specifies the Segoe MDL2 Assets glyph to use.
+        /// </returns>
+        public Symbol Symbol
+        {
+            get => (Symbol)GetValue(SymbolProperty);
+            set => SetValue(SymbolProperty, value);
+        }
+    }
 }
